vk-1940-codable-classes: Adopted Codable to progress classes, added tests
