+++ MapboxCoreNavigation/RouteProgress.swift
@@ -5,7 +5,7 @@ import Turf
 /**
  `RouteProgress` stores the user’s progress along a route.
  */
-open class RouteProgress {
+open class RouteProgress: Codable {
     private static let reroutingAccuracy: CLLocationAccuracy = 90
     
     var indexedRoute: IndexedRoute
@@ -296,401 +296,68 @@ open class RouteProgress {
         
         calculateLegsCongestion()
     }
-}
-
-/**
- `RouteLegProgress` stores the user’s progress along a route leg.
- */
-open class RouteLegProgress {
-    /**
-     Returns the current `RouteLeg`.
-     */
-    public let leg: RouteLeg
-
-    /**
-     Index representing the current step.
-     */
-    public var stepIndex: Int {
-        didSet {
-            assert(stepIndex >= 0 && stepIndex < leg.steps.endIndex)
-            currentStepProgress = RouteStepProgress(step: currentStep)
-        }
-    }
-
-    /**
-     The remaining steps for user to complete.
-     */
-    public var remainingSteps: [RouteStep] {
-        return Array(leg.steps.suffix(from: stepIndex + 1))
-    }
-
-    /**
-     Total distance traveled in meters along current leg.
-     */
-    public var distanceTraveled: CLLocationDistance {
-        return leg.steps.prefix(upTo: stepIndex).map { $0.distance }.reduce(0, +) + currentStepProgress.distanceTraveled
-    }
-
-    /**
-     Duration remaining in seconds on current leg.
-     */
-    public var durationRemaining: TimeInterval {
-        return remainingSteps.map { $0.expectedTravelTime }.reduce(0, +) + currentStepProgress.durationRemaining
-    }
-
-    /**
-     Distance remaining on the current leg.
-     */
-    public var distanceRemaining: CLLocationDistance {
-        return remainingSteps.map { $0.distance }.reduce(0, +) + currentStepProgress.distanceRemaining
-    }
-
-    /**
-     Number between 0 and 1 representing how far along the current leg the user has traveled.
-     */
-    public var fractionTraveled: Double {
-        return distanceTraveled / leg.distance
-    }
-
-    public var userHasArrivedAtWaypoint = false
-
-    /**
-     Returns the `RouteStep` before a given step. Returns `nil` if there is no step prior.
-     */
-    public func stepBefore(_ step: RouteStep) -> RouteStep? {
-        guard let index = leg.steps.firstIndex(of: step) else {
-            return nil
-        }
-        if index > 0 {
-            return leg.steps[index-1]
-        }
-        return nil
-    }
-
-    /**
-     Returns the `RouteStep` after a given step. Returns `nil` if there is not a step after.
-     */
-    public func stepAfter(_ step: RouteStep) -> RouteStep? {
-        guard let index = leg.steps.firstIndex(of: step) else {
-            return nil
-        }
-        if index+1 < leg.steps.endIndex {
-            return leg.steps[index+1]
-        }
-        return nil
-    }
-
-    /**
-     Returns the `RouteStep` before the current step.
-
-     If there is no `priorStep`, nil is returned.
-     */
-    public var priorStep: RouteStep? {
-        guard stepIndex - 1 >= 0 else {
-            return nil
-        }
-        return leg.steps[stepIndex - 1]
-    }
-
-    /**
-     Returns the current `RouteStep` for the leg the user is on.
-     */
-    public var currentStep: RouteStep {
-        return leg.steps[stepIndex]
-    }
-
-    /**
-     Returns the upcoming `RouteStep`.
-
-     If there is no `upcomingStep`, nil is returned.
-     */
-    @available(swift, obsoleted: 0.1, renamed: "upcomingStep")
-    public var upComingStep: RouteStep? {
-        fatalError()
+    
+    // MARK: - Codable implementation
+    
+    private enum CodingKeys: String, CodingKey {
+        case indexedRoute
+        case indexedRouteIndex
+        case routeOptions
+        case legIndex
+        case currentLegProgress
+        case congestionTravelTimesSegmentsByStep
+        case congestionTimesPerStep
     }
     
-    public var upcomingStep: RouteStep? {
-        guard stepIndex + 1 < leg.steps.endIndex else {
-            return nil
-        }
-        return leg.steps[stepIndex + 1]
+    private struct TimedCongestionLevelCodable: Codable {
+        var congestionLevel: CongestionLevel
+        var travelTime: TimeInterval
     }
-
-    /**
-     Returns step 2 steps ahead.
-
-     If there is no `followOnStep`, nil is returned.
-     */
-    public var followOnStep: RouteStep? {
-        guard stepIndex + 2 < leg.steps.endIndex else {
-            return nil
-        }
-        return leg.steps[stepIndex + 2]
-    }
-
-    /**
-     Return bool whether step provided is the current `RouteStep` the user is on.
-     */
-    public func isCurrentStep(_ step: RouteStep) -> Bool {
-        return step == currentStep
-    }
-
-    /**
-     Returns the progress along the current `RouteStep`.
-     */
-    public var currentStepProgress: RouteStepProgress
-
-    /**
-     Intializes a new `RouteLegProgress`.
-
-     - parameter leg: Leg on a `Route`.
-     - parameter stepIndex: Current step the user is on.
-     */
-    public init(leg: RouteLeg, stepIndex: Int = 0, spokenInstructionIndex: Int = 0) {
-        self.leg = leg
-        self.stepIndex = stepIndex
-        
-        precondition(leg.steps.indices.contains(stepIndex), "It's not possible to create RouteLegProgress without any steps or when stepIndex is higher than steps count.")
-        
-        currentStepProgress = RouteStepProgress(step: leg.steps[stepIndex], spokenInstructionIndex: spokenInstructionIndex)
-    }
-
-    typealias StepIndexDistance = (index: Int, distance: CLLocationDistance)
-
-    func closestStep(to coordinate: CLLocationCoordinate2D) -> StepIndexDistance? {
-        var currentClosest: StepIndexDistance?
-        let remainingSteps = leg.steps.suffix(from: stepIndex)
-
-        for (currentStepIndex, step) in remainingSteps.enumerated() {
-            guard let shape = step.shape else { continue }
-            guard let closestCoordOnStep = shape.closestCoordinate(to: coordinate) else { continue }
-            let closesCoordOnStepDistance = closestCoordOnStep.coordinate.distance(to: coordinate)
-            let foundIndex = currentStepIndex + stepIndex
-
-            // First time around, currentClosest will be `nil`.
-            guard let currentClosestDistance = currentClosest?.distance else {
-                currentClosest = (index: foundIndex, distance: closesCoordOnStepDistance)
-                continue
-            }
-
-            if closesCoordOnStepDistance < currentClosestDistance {
-                currentClosest = (index: foundIndex, distance: closesCoordOnStepDistance)
+    
+    private func decodeCongestion(_ decodable: [[[TimedCongestionLevelCodable]]]) -> [[[TimedCongestionLevel]]] {
+        return decodable.map {
+            return $0.map {
+                return $0.map {
+                    return ($0.congestionLevel, $0.travelTime)
+                }
             }
         }
-
-        return currentClosest
     }
     
-    /**
-     The waypoints remaining on the current leg, not including the leg’s destination.
-     */
-    func remainingWaypoints(among waypoints: [Waypoint]) -> [Waypoint] {
-        guard waypoints.count > 1 else {
-            // The leg has only a source and no via points. Save ourselves a call to RouteLeg.coordinates, which can be expensive.
-            return []
-        }
-        let legPolyline = leg.shape
-        guard let userCoordinateIndex = legPolyline.indexedCoordinateFromStart(distance: distanceTraveled)?.index else {
-            // The leg is empty, so none of the waypoints are meaningful.
-            return []
-        }
-        var slice = legPolyline
-        var accumulatedCoordinates = 0
-        return Array(waypoints.drop { (waypoint) -> Bool in
-            let newSlice = slice.sliced(from: waypoint.coordinate)!
-            accumulatedCoordinates += slice.coordinates.count - newSlice.coordinates.count
-            slice = newSlice
-            return accumulatedCoordinates <= userCoordinateIndex
-        })
-    }
-
-    /**
-     Returns the SpeedLimit for the current position along the route. Returns SpeedLimit.invalid if the speed limit is unknown or missing.
-     
-     The maximum speed may be an advisory speed limit for segments where legal limits are not posted, such as highway entrance and exit ramps. If the speed limit along a particular segment is unknown, it is set to `nil`. If the speed is unregulated along the segment, such as on the German _Autobahn_ system, it is represented by a measurement whose value is `Double.infinity`.
-     
-     Speed limit data is available in [a number of countries and territories worldwide](https://docs.mapbox.com/help/how-mapbox-works/directions/).
-     */
-    public var currentSpeedLimit: Measurement<UnitSpeed>? {
-        guard let segmentMaximumSpeedLimits = leg.segmentMaximumSpeedLimits else {
-            return nil
-        }
-        
-        let distanceTraveled = currentStepProgress.distanceTraveled
-        guard var index = currentStep.shape?.indexedCoordinateFromStart(distance: distanceTraveled)?.index else {
-            return nil
-        }
-        
-        var range = leg.segmentRangesByStep[stepIndex]
-        
-        // indexedCoordinateFromStart(distance:) can return a coordinate indexed to the last coordinate of the step, which is past any segment on the current step.
-        if index == range.count && upcomingStep != nil {
-            range = leg.segmentRangesByStep[stepIndex.advanced(by: 1)]
-            index = 0
-        }
-        guard index < range.count && range.upperBound <= segmentMaximumSpeedLimits.endIndex else {
-            return nil
-        }
-        
-        let speedLimit = segmentMaximumSpeedLimits[range][range.lowerBound.advanced(by: index)]
-        if let speedUnit = currentStep.speedLimitUnit {
-            return speedLimit?.converted(to: speedUnit)
-        } else {
-            return speedLimit
+    private func encodeCongestion(_ encodable: [[[TimedCongestionLevel]]]) -> [[[TimedCongestionLevelCodable]]] {
+        return encodable.map {
+            return $0.map {
+                return $0.map {
+                    return TimedCongestionLevelCodable(congestionLevel: $0.0,
+                                                       travelTime: $0.1)
+                }
+            }
         }
     }
-}
-
-/**
- `RouteStepProgress` stores the user’s progress along a route step.
- */
-open class RouteStepProgress {
-    /**
-     Returns the current `RouteStep`.
-     */
-    public let step: RouteStep
-
-    /**
-     Returns distance user has traveled along current step.
-     */
-    public var distanceTraveled: CLLocationDistance = 0
-
-    /**
-     Returns distance from user to end of step.
-     */
-    public var userDistanceToManeuverLocation: CLLocationDistance
-
-    /**
-     Total distance in meters remaining on current step.
-     */
-    public var distanceRemaining: CLLocationDistance {
-        return step.distance - distanceTraveled
-    }
-
-    /**
-     Number between 0 and 1 representing fraction of current step traveled.
-     */
-    public var fractionTraveled: Double {
-        guard step.distance > 0 else { return 1 }
-        return distanceTraveled / step.distance
-    }
-
-    /**
-     Number of seconds remaining on current step.
-     */
-    public var durationRemaining: TimeInterval {
-        return (1 - fractionTraveled) * step.expectedTravelTime
+        
+    required public init(from decoder: Decoder) throws {
+        let container = try decoder.container(keyedBy: CodingKeys.self)
+        
+        let route = try container.decode(Route.self, forKey: .indexedRoute)
+        let routeIndex = try container.decode(Int.self, forKey: .indexedRouteIndex)
+        self.indexedRoute = (route, routeIndex)
+        self.routeOptions = try container.decode(RouteOptions.self, forKey: .routeOptions)
+        self.legIndex = try container.decode(Int.self, forKey: .legIndex)
+        self.currentLegProgress = try container.decode(RouteLegProgress.self, forKey: .currentLegProgress)
+        self.congestionTravelTimesSegmentsByStep = decodeCongestion(try container.decode([[[TimedCongestionLevelCodable]]].self,
+                                                                                         forKey: .congestionTravelTimesSegmentsByStep))
+        self.congestionTimesPerStep = try container.decode([[[CongestionLevel: TimeInterval]]].self, forKey: .congestionTimesPerStep)
     }
     
-    /**
-     Intializes a new `RouteStepProgress`.
-
-     - parameter step: Step on a `RouteLeg`.
-     */
-    public init(step: RouteStep, spokenInstructionIndex: Int = 0) {
-        self.step = step
-        self.userDistanceToManeuverLocation = step.distance
-        self.intersectionIndex = 0
-        self.spokenInstructionIndex = spokenInstructionIndex
-    }
-
-    /**
-     All intersections on the current `RouteStep` and also the first intersection on the upcoming `RouteStep`.
-
-     The upcoming `RouteStep` first `Intersection` is added because it is omitted from the current step.
-     */
-    public var intersectionsIncludingUpcomingManeuverIntersection: [Intersection]?
-
-    /**
-     The next intersection the user will travel through.
-
-     The step must contain `intersectionsIncludingUpcomingManeuverIntersection` otherwise this property will be `nil`.
-     */
-    public var upcomingIntersection: Intersection? {
-        guard let intersections = intersectionsIncludingUpcomingManeuverIntersection, intersections.startIndex..<intersections.endIndex-1 ~= intersectionIndex else {
-            return nil
-        }
-
-        return intersections[intersections.index(after: intersectionIndex)]
-    }
-
-    /**
-     Index representing the current intersection.
-     */
-    public var intersectionIndex: Int = 0
-
-    /**
-     The current intersection the user will travel through.
-
-     The step must contain `intersectionsIncludingUpcomingManeuverIntersection` otherwise this property will be `nil`.
-     */
-    public var currentIntersection: Intersection? {
-        guard let intersections = intersectionsIncludingUpcomingManeuverIntersection, intersections.startIndex..<intersections.endIndex ~= intersectionIndex else {
-            return nil
-        }
-
-        return intersections[intersectionIndex]
-    }
-
-    /**
-     Returns an array of the calculated distances from the current intersection to the next intersection on the current step.
-     */
-    public var intersectionDistances: Array<CLLocationDistance>?
-
-    /**
-     The distance in meters the user is to the next intersection they will pass through.
-     */
-    public var userDistanceToUpcomingIntersection: CLLocationDistance?
-
-    /**
-     Index into `step.instructionsDisplayedAlongStep` representing the current visual instruction for the step.
-     */
-    public var visualInstructionIndex: Int = 0
-
-    /**
-     An `Array` of remaining `VisualInstruction` for a step.
-     */
-    public var remainingVisualInstructions: [VisualInstructionBanner]? {
-        guard let visualInstructions = step.instructionsDisplayedAlongStep else { return nil }
-        return Array(visualInstructions.suffix(from: visualInstructionIndex))
-    }
-
-    /**
-     Index into `step.instructionsSpokenAlongStep` representing the current spoken instruction.
-     */
-    public var spokenInstructionIndex: Int = 0
-
-    /**
-     An `Array` of remaining `SpokenInstruction` for a step.
-     */
-    public var remainingSpokenInstructions: [SpokenInstruction]? {
-        guard let instructions = step.instructionsSpokenAlongStep else { return nil }
-        return Array(instructions.suffix(from: spokenInstructionIndex))
-    }
-
-    /**
-     Current spoken instruction for the user's progress along a step.
-     */
-    public var currentSpokenInstruction: SpokenInstruction? {
-        guard let instructionsSpokenAlongStep = step.instructionsSpokenAlongStep else { return nil }
-        guard spokenInstructionIndex < instructionsSpokenAlongStep.count else { return nil }
-        return instructionsSpokenAlongStep[spokenInstructionIndex]
-    }
-
-    /**
-     Current visual instruction for the user's progress along a step.
-     */
-    public var currentVisualInstruction: VisualInstructionBanner? {
-        guard let instructionsDisplayedAlongStep = step.instructionsDisplayedAlongStep else { return nil }
-        guard visualInstructionIndex < instructionsDisplayedAlongStep.count else { return nil }
-        return instructionsDisplayedAlongStep[visualInstructionIndex]
-    }
-    
-    public var keyPathsAffectingValueForRemainingVisualInstructions: Set<String> {
-        return ["step.instructionsDisplayedAlongStep", "visualInstructionIndex"]
-    }
-    
-    public var keyPathsAffectingValueForRemainingSpokenInstructions: Set<String> {
-        return ["step.instructionsDisplayedAlongStep", "spokenInstructionIndex"]
+    public func encode(to encoder: Encoder) throws {
+        var container = encoder.container(keyedBy: CodingKeys.self)
+        
+        try container.encode(indexedRoute.0, forKey: .indexedRoute)
+        try container.encode(indexedRoute.1, forKey: .indexedRouteIndex)
+        try container.encode(routeOptions, forKey: .routeOptions)
+        try container.encode(legIndex, forKey: .legIndex)
+        try container.encode(currentLegProgress, forKey: .currentLegProgress)
+        try container.encode(encodeCongestion(congestionTravelTimesSegmentsByStep), forKey: .congestionTravelTimesSegmentsByStep)
+        try container.encode(congestionTimesPerStep, forKey: .congestionTimesPerStep)
     }
 }
